sr(api): move TrainingForAdmin model in devcomp
