+++ api/src/devcomp/domain/read-models/TrainingForAdmin.js
